Initial commit
This actually is my code from months ago that exported my data from
Rdio, but that's no longer necessary (or useful) because Rdio has shut
down and made zips available of users' bulk data. I'm going to remove
this file in the next commit, but I wanted to commit it first for
posterity's sake.